add "debug mode" and fix some bugs

+++ src/preevaluation.ts
@@ -1,15 +1,15 @@
-import * as ts from 'typescript';
-import * as path from 'path';
-import * as fs from 'fs';
-import * as deleteDir from 'rimraf';
 import * as shell from 'child_process';
-import * as findCacheDir from 'find-cache-dir';
-import prevalOptions from './prevalOptions';
+import crypto from 'crypto';
+import findCacheDir = require('find-cache-dir');
+import * as fs from 'fs';
+import * as path from 'path';
+import * as deleteDir from 'rimraf';
+import * as ts from 'typescript';
+import PrevalOptions from './prevalOptions';
 
-const sha1 = require('sha1');
-
-let _options: prevalOptions = {
+let OPTIONS: PrevalOptions = {
   cacheActivated: false,
+  debug: false,
   mode: 'development'
 };
 
@@ -32,23 +32,31 @@ function preevaluationTransformer<T extends ts.Node>(): ts.TransformerFactory<
 
       if (
         ts.isTaggedTemplateExpression(node) &&
-        node.tag.getText() == tagName
+        node.tag.getText() === tagName
       ) {
         let code: string = node.template.getText().slice(1, -1);
 
         const dir = path.dirname(node.getSourceFile().fileName);
 
-        const hash = sha1(code);
+        const hash = crypto
+          .createHash('sha1')
+          .update(code)
+          .digest('hex');
 
-        const evalFileDir = dir.split('/')[dir.split('/').length - 1];
+        // const evalFileDir = dir.split('/')[dir.split('/').length - 1];
         const evalFilename = '_eval.' + hash;
         const evalPath = path.join(dir, evalFilename + '.ts');
         const evalDir = `${dir}/_temp_${hash}`;
-        const evalResult =
-          findCacheDir({
-            name: `ts-transformer-preval/${_options.mode}`,
-            create: true
-          }) + `/${hash}.json`;
+        const cacheDir = findCacheDir({
+          create: true,
+          name: `ts-transformer-preval-macro/${OPTIONS.mode}`
+        });
+
+        if (cacheDir === null) {
+          throw new Error('not able to create a cache dir');
+        }
+
+        const evalResult = path.join(cacheDir, `/${hash}.json`);
 
         code = `${code} 
 
@@ -56,7 +64,7 @@ run().then((result) => {
 console.log(JSON.stringify(result));
 });`;
 
-        if (fs.existsSync(evalResult) && _options.cacheActivated) {
+        if (fs.existsSync(evalResult) && OPTIONS.cacheActivated) {
           const cachedRawResult = fs.readFileSync(evalResult);
 
           return transformResult(cachedRawResult);
@@ -68,24 +76,30 @@ console.log(JSON.stringify(result));
           shell.execSync(`tsc ${evalPath} --outDir ${evalDir}`);
         } catch (e) {
           // TODO parse tsc errors and decide if something gone really wrong
-          //fs.unlinkSync(evalPath);
-          //deleteDir.sync(evalDir);
-          //throw e;
+          if (!OPTIONS.debug) {
+            fs.unlinkSync(evalPath);
+            deleteDir.sync(evalDir);
+            throw e;
+          }
         }
 
         let rawResult = 'null';
         try {
           rawResult = shell
-            .execSync(`node ${evalDir}/${evalFileDir}/${evalFilename}.js`)
+            .execSync(`node ${evalDir}/${evalFilename}.js`)
             .toString();
         } catch (e) {
-          fs.unlinkSync(evalPath);
-          deleteDir.sync(evalDir);
-          throw e;
+          if (!OPTIONS.debug) {
+            fs.unlinkSync(evalPath);
+            deleteDir.sync(evalDir);
+            throw e;
+          }
         }
 
-        fs.unlinkSync(evalPath);
-        deleteDir.sync(evalDir);
+        if (!OPTIONS.debug) {
+          fs.unlinkSync(evalPath);
+          deleteDir.sync(evalDir);
+        }
 
         fs.writeFileSync(evalResult, rawResult);
 
@@ -113,9 +127,9 @@ function transformResult(rawResult: any) {
   );
 }
 
-export default (options: prevalOptions) => {
-  _options = {
-    ..._options,
+export default (options: PrevalOptions) => {
+  OPTIONS = {
+    ...OPTIONS,
     ...options
   };
